Feature ETP-435: Add drawer Utils
