+++ packages/ComponentLibrary/src/components/Drawer/DrawerSection/index.tsx
@@ -11,23 +11,7 @@ import MenuTitle from '../MenuTitle';
 import { theme } from '../../../theme';
 import { DrawerSectionProps } from '../types';
 import { useParams } from 'react-router-dom';
-import { Menu } from '@workspaceui/etendohookbinder/src/api/types';
-
-const findActive = (
-  windowId: string | undefined,
-  items: Menu[] | undefined = [],
-): boolean => {
-  if (!items || !windowId) return false;
-  const stack: Menu[] = [...items];
-  while (stack.length > 0) {
-    const item = stack.pop();
-    if (item) {
-      if (item.windowId === windowId) return true;
-      if (item.children) stack.push(...item.children);
-    }
-  }
-  return false;
-};
+import { findActive } from '../../../utils/drawerUtils';
 
 const DrawerSection: React.FC<DrawerSectionProps> = ({
   item,
